# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: سند – منصة الخدمات والمهنيين في اليمن
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /providers:
    get:
      operationId: listProviders
      tags: [providers]
      summary: Search and list service providers
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: city
          in: query
          schema:
            type: ["string", "null"]
        - name: district
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: minRating
          in: query
          schema:
            type: ["number", "null"]
        - name: lat
          in: query
          schema:
            type: ["number", "null"]
        - name: lng
          in: query
          schema:
            type: ["number", "null"]
        - name: sortBy
          in: query
          schema:
            type: ["string", "null"]
            enum: [distance, rating, experience, null]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of providers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvidersPage"
  /providers/{id}:
    get:
      operationId: getProvider
      tags: [providers]
      summary: Get provider profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderDetail"
    patch:
      operationId: updateProvider
      tags: [providers]
      summary: Update provider profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderDetail"
  /providers/nearby:
    get:
      operationId: getNearbyProviders
      tags: [providers]
      summary: Get nearest providers by location
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lng
          in: query
          required: true
          schema:
            type: number
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: radiusKm
          in: query
          schema:
            type: ["number", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Nearby providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProviderSummary"
  /providers/top-rated:
    get:
      operationId: getTopRatedProviders
      tags: [providers]
      summary: Get top rated providers
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Top rated providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProviderSummary"
  /providers/most-requested:
    get:
      operationId: getMostRequestedProviders
      tags: [providers]
      summary: Get most requested providers
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Most requested providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProviderSummary"
  /providers/{id}/portfolio:
    get:
      operationId: getProviderPortfolio
      tags: [providers]
      summary: Get provider portfolio images
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Portfolio items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioItem"
    post:
      operationId: addPortfolioItem
      tags: [providers]
      summary: Add portfolio item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioItem"

  # ─── SERVICE REQUESTS ────────────────────────────────────────────────────────
...
  /home-feed:
    get:
      operationId: getHomeFeed
      tags: [providers]
      summary: Home page aggregated feed
      parameters:
        - name: lat
          in: query
          schema:
            type: ["number", "null"]
        - name: lng
          in: query
          schema:
            type: ["number", "null"]
      responses:
        "200":
          description: Home feed data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeFeed"
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, icon, providerCount]
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
        providerCount:
          type: integer
...
    ProviderSummary:
      type: object
      required: [id, name, categoryName, city, district, rating, reviewCount, completedJobs, yearsExperience, isVerified, isAvailable]
      properties:
        id:
          type: integer
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        categoryName:
          type: string
        categoryIcon:
          type: ["string", "null"]
        city:
          type: string
        district:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        completedJobs:
          type: integer
        yearsExperience:
          type: integer
        hourlyRate:
          type: ["number", "null"]
        isVerified:
          type: boolean
        isAvailable:
          type: boolean
        distanceKm:
          type: ["number", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
    ProviderDetail:
      type: object
      required: [id, name, categoryName, city, district, rating, reviewCount, completedJobs, yearsExperience, isVerified, isAvailable, bio]
      properties:
        id:
          type: integer
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: string
        categoryIcon:
          type: ["string", "null"]
        city:
          type: string
        district:
          type: string
        bio:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        completedJobs:
          type: integer
        yearsExperience:
          type: integer
        hourlyRate:
          type: ["number", "null"]
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        isVerified:
          type: boolean
        isAvailable:
          type: boolean
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        isFavorited:
          type: boolean
        createdAt:
          type: string
    ProviderUpdate:
      type: object
      properties:
        bio:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        district:
          type: ["string", "null"]
        yearsExperience:
          type: ["integer", "null"]
        hourlyRate:
          type: ["number", "null"]
        whatsapp:
          type: ["string", "null"]
        isAvailable:
          type: ["boolean", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
    ProvidersPage:
      type: object
      required: [providers, total, page, limit]
      properties:
        providers:
          type: array
          items:
            $ref: "#/components/schemas/ProviderSummary"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    PortfolioItem:
      type: object
      required: [id, imageUrl, providerId]
      properties:
        id:
          type: integer
        imageUrl:
          type: string
        description:
          type: ["string", "null"]
        providerId:
          type: integer
        createdAt:
          type: string
    PortfolioInput:
      type: object
      required: [imageUrl]
      properties:
        imageUrl:
          type: string
        description:
          type: ["string", "null"]

    # ── SERVICE REQUEST ──────────────────────────────────────────────────────────
    ServiceRequest:
      type: object
      required: [id, clientId, providerId, status, serviceType, description, city, district, scheduledAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        clientAvatarUrl:
          type: ["string", "null"]
        providerId:
          type: integer
        providerName:
          type: string
        providerAvatarUrl:
          type: ["string", "null"]
        providerCategoryName:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected, in_progress, completed, cancelled]
        serviceType:
          type: string
        description:
          type: string
        city:
          type: string
        district:
          type: string
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        scheduledAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        isImmediate:
          type: boolean
        createdAt:
          type: string
...
    HomeFeed:
      type: object
      required: [categories, nearbyProviders, topRatedProviders, mostRequestedProviders, recentRequests]
      properties:
        categories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        nearbyProviders:
          type: array
          items:
            $ref: "#/components/schemas/ProviderSummary"
        topRatedProviders:
          type: array
          items:
            $ref: "#/components/schemas/ProviderSummary"
        mostRequestedProviders:
          type: array
          items:
            $ref: "#/components/schemas/ProviderSummary"
        recentRequests:
          type: array
          items:
            $ref: "#/components/schemas/ServiceRequest"