# lib/api-zod/src/generated/types/providersPage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * سند – منصة الخدمات والمهنيين في اليمن
 * OpenAPI spec version: 0.1.0
 */
import type { ProviderSummary } from './providerSummary';

export interface ProvidersPage {
  providers: ProviderSummary[];
  total: number;
  page: number;
  limit: number;
}